+++ core/routes/versions.ts
@@ -7,6 +7,7 @@ import {
   createVersion,
   deleteVersion,
   publishVersion,
+  saveVersion,
   syncVersion,
   updateVersion,
 } from "../usecases/versions.ts";
@@ -27,44 +28,51 @@ app.post("/*", async ({ request, cms, user, next }) => {
   const body = await request.formData();
   const name = body.get("name") as string;
 
-  const response = new Response(null, {
-    status: 302,
-    headers: new Headers({
-      "Location": getPath(basePath),
-      "X-Lume-CMS": "reload",
-    }),
-  });
+  function redirect(reload = true) {
+    return new Response(null, {
+      status: 302,
+      headers: new Headers({
+        "Location": getPath(basePath),
+        "X-Lume-CMS": reload ? "reload" : "",
+      }),
+    });
+  }
 
   return next()
     /* POST /versions/create */
     .post("/create", () => {
       createVersion(user, git, name);
-      return response;
+      return redirect();
+    })
+    /* POST /versions/save */
+    .post("/save", () => {
+      saveVersion(user, git, name);
+      return redirect(false);
     })
     /* POST /versions/change */
     .post("/change", () => {
       changeVersion(user, git, name);
-      return response;
+      return redirect();
     })
     /* POST /versions/publish */
     .post("/publish", () => {
       publishVersion(user, git, name);
-      return response;
+      return redirect();
     })
     /* POST /versions/delete */
     .post("/delete", () => {
       deleteVersion(user, git, name);
-      return response;
+      return redirect();
     })
     /* POST /versions/update */
     .post("/update", () => {
       updateVersion(git, name);
-      return response;
+      return redirect();
     })
     /* POST /versions/sync */
     .post("/sync", () => {
       syncVersion(user, git, name);
-      return response;
+      return redirect();
     });
 });