+++ core/usecases/versions.ts
@@ -2,8 +2,7 @@ import type Git from "../git.ts";
 import type User from "../user.ts";
 
 export function createVersion(user: User, git: Git, name: string) {
-  git.create(name);
-  git.change(user, name);
+  git.create(user, name);
 }
 
 export function changeVersion(user: User, git: Git, name: string) {
@@ -15,6 +14,13 @@ export function syncVersion(user: User, git: Git, name: string) {
     throw new Error("The current version doesn't match");
   }
   git.sync(user);
+}
+
+export function saveVersion(user: User, git: Git, name: string) {
+  if (git.current().name !== name) {
+    throw new Error("The current version doesn't match");
+  }
+  git.save(user);
 }
 
 export function updateVersion(git: Git, name: string) {